+++ python/fnllm/fnllm/limiting/__init__.py
@@ -2,13 +2,13 @@
 
 """Limiting base package."""
 
-from .base import Limiter, Manifest
+from .base import Limiter
 from .composite import CompositeLimiter
 from .concurrency import ConcurrencyLimiter
 from .noop_llm import NoopLimiter
 from .rpm import RPMLimiter
 from .tpm import TPMLimiter
-from .types import LimitReconciler
+from .types import LimitReconciler, Manifest, Reconciliation
 
 __all__ = [
     "CompositeLimiter",
@@ -18,5 +18,6 @@ __all__ = [
     "Manifest",
     "NoopLimiter",
     "RPMLimiter",
+    "Reconciliation",
     "TPMLimiter",
 ]
